fix: add POST /api/v1/shops endpoint for creating shops and fix ChatServiceImpl type error
- Created ShopDtos with CreateShopRequest and ShopResponse DTOs
- Updated ShopService interface with createShop method
- Implemented createShop in ShopServiceImpl with proper validation
- Added POST /api/v1/shops endpoint in ShopController (placed before /{id} to avoid routing conflicts)
- Fixed ChatServiceImpl type conversion error (long to Integer)

This resolves the 500 error when sellers try to create a shop.

+++ src/main/java/kg/bishkek/fucent/fusent/service/impl/ChatServiceImpl.java
@@ -142,7 +142,7 @@ public class ChatServiceImpl implements ChatService {
             otherUser.getEmail(),
             firstMessage.getMessageText(),
             firstMessage.getCreatedAt(),
-            unreadCount
+            (int) unreadCount
         );
     }
